// src/types/Character.ts
export type Character = {
	health: number
	maxHealth: number
	name: string
	baseDamage: number
	randomDamage: number
	baseHeal: number
	randomHeal: number
	healCount: number
	fireballBaseDamage: number
	fireballRandomDamage: number
	fireballCount: number
	critChance: number
	critFactor: number
	emoji: string
}